+++ src/constants.ts
@@ -6,6 +6,7 @@ import {
   GERMANY_COUNTRY,
   PORTUGAL_COUNTRY,
   SPAIN_COUNTRY,
+  SWEDEN_COUNTRY,
   SWITZERLAND_COUNTRY,
 } from "./geo/countries.js";
 
@@ -16,6 +17,7 @@ export const DENMARK_CENTER: [number, number] = DENMARK_COUNTRY.center;
 export const SWITZERLAND_CENTER: [number, number] = SWITZERLAND_COUNTRY.center;
 export const PORTUGAL_CENTER: [number, number] = PORTUGAL_COUNTRY.center;
 export const AUSTRIA_CENTER: [number, number] = AUSTRIA_COUNTRY.center;
+export const SWEDEN_CENTER: [number, number] = SWEDEN_COUNTRY.center;
 export const CZECHIA_CENTER: [number, number] = COUNTRIES.CZ.center;
 export const POLAND_CENTER: [number, number] = COUNTRIES.PL.center;
 export const DEFAULT_ZOOM = 6;